+++ src/async-consume-many.d.ts
@@ -1,0 +1,2 @@
+import { asyncConsumeMany } from './index'
+export default asyncConsumeMany