+++ src/async-consume.mjs
@@ -2,8 +2,9 @@ import ensureAsyncIterable from './internal/ensure-async-iterable'
 import curry from './internal/curry'
 
 async function asyncConsume (func, iterable) {
+  let c = 0
   for await (const item of ensureAsyncIterable(iterable)) {
-    await func(item)
+    await func(item, c++)
   }
 }