+++ src/consume-many.d.ts
@@ -1,0 +1,2 @@
+import { consumeMany } from './index'
+export default consumeMany